requirements.txt
ansicon==1.89.0
argcomplete==3.6.2
awsebcli==3.23.2
bcrypt==4.2.1
blessed==1.20.0
blinker==1.9.0
boto3==1.38.1
botocore==1.38.1
cement==2.10.14
certifi==2025.1.31
cffi==1.17.1
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
dnspython==2.7.0
docx==0.2.4
durationpy==0.9
ecdsa==0.19.1
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==21.2.0
hjson==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jinxed==1.3.0
jmespath==1.0.1
kappa==0.6.0
linecache2==1.0.0
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
placebo==0.9.0
platformdirs==4.3.7
pyasn1==0.6.1
pycparser==2.22
pymongo==4.6.1
# pypiwin32==223  # Windows-specific package - commented out
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
Python_IO==0.3
python-jose==3.3.0
python-pptx==1.0.2
python-slugify==8.0.4
# pythonnet==3.0.5  # May cause issues on Linux - commented out
pytz==2024.2
# pywin32==310  # Windows-specific package - commented out
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
s3transfer==0.12.0
semantic-version==2.10.0
six==1.17.0
termcolor==2.5.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
traceback2==1.4.0
troposphere==4.9.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
userpath==1.9.2
wcwidth==0.2.13
Werkzeug==3.1.3
XlsxWriter==3.2.2
zappa==0.59.0
zipfile36==0.1.3